refactor recoGenRecommander, separate recoGen and publicFeedGen to make classes smaller and easy to maintain.
