+++ kifi-backend/modules/curator/test/com/keepit/curator/RecommendationGenerationCommanderTest.scala
@@ -10,7 +10,7 @@ import com.keepit.common.healthcheck.FakeHealthcheckModule
 import com.keepit.curator.model._
 import com.keepit.graph.FakeGraphServiceModule
 import com.keepit.heimdal.FakeHeimdalServiceClientModule
-import com.keepit.model.{ UriRecommendationScores, User, NormalizedURI }
+import com.keepit.model.{ User, NormalizedURI }
 import com.keepit.search.FakeSearchServiceClientModule
 import org.specs2.mutable.Specification
 
@@ -77,127 +77,54 @@ class RecommendationGenerationCommanderTest extends Specification with CuratorTe
     Seq(rec1, rec2, rec3)
   }
 
-  def setupFeeds(): Seq[PublicFeed] = {
-    val rec1 = makePublicFeed(1, 3.0f)
-    val rec2 = makePublicFeed(2, 0.99f)
-    val rec3 = makePublicFeed(3, 0.5f)
-    Seq(rec1, rec2, rec3)
+  "pre-compute recommendations" in {
+    withDb(modules: _*) { implicit injector =>
+      val (user1, user2, shoebox) = setup()
+      val user1Keeps = makeKeeps(user1, 5, shoebox)
+      val user2Keeps = makeKeeps(user2, 5, shoebox)
+      val seedCommander = inject[SeedIngestionCommander]
+      val seedItemRepo = inject[RawSeedItemRepo]
+      val commander = inject[RecommendationGenerationCommander]
+      Await.result(seedCommander.ingestAllKeeps(), Duration(10, "seconds"))
+      db.readWrite { implicit session => seedItemRepo.assignSequenceNumbers(1000) }
+      commander.precomputeRecommendations()
+      val futUnit = commander.precomputeRecommendations()
+      Await.result(futUnit, Duration(10, "seconds"))
+      val result = commander.getTopRecommendations(Id[User](42), 1)
+      val recs = Await.result(result, Duration(10, "seconds"))
+      recs.size === 0
+    }
   }
 
-  "RecommendationGenerationCommanderTest" should {
-    "get public feeds" in {
-      withDb(modules: _*) {
-        implicit injector =>
-          val repo = inject[PublicFeedRepo]
-          db.readWrite {
-            implicit s =>
-              val feeds = setupFeeds()
-              repo.save(feeds(0))
-              repo.save(feeds(1))
-              repo.save(feeds(2))
-          }
+  "re-compute recos for user" in {
+    withDb(modules: _*) { implicit injector =>
 
-          val commander = inject[RecommendationGenerationCommander]
+      val (user1, user2, shoebox) = setup()
+      val user1Keeps = makeKeeps(user1, 5, shoebox)
+      val user2Keeps = makeKeeps(user2, 5, shoebox)
+      val seedCommander = inject[SeedIngestionCommander]
+      val seedItemRepo = inject[RawSeedItemRepo]
+      val commander = inject[RecommendationGenerationCommander]
+      Await.result(seedCommander.ingestAllKeeps(), Duration(10, "seconds"))
+      db.readWrite { implicit session => seedItemRepo.assignSequenceNumbers(1000) }
+      commander.precomputeRecommendations()
+      val futUnit = commander.precomputeRecommendations()
+      Await.result(futUnit, Duration(10, "seconds"))
+      val result = commander.getTopRecommendations(Id[User](42), 1)
+      val recs = Await.result(result, Duration(10, "seconds"))
+      recs.size === 0
 
-          val result = commander.getPublicFeeds(2)
-          val feeds = Await.result(result, Duration(10, "seconds"))
-          feeds.size === 2
-          feeds(0).uriId === Id[NormalizedURI](1)
-          feeds(0).publicMasterScore === 3.0f
+      val resetFut = commander.resetUser(Id[User](42))
+      Await.result(resetFut, Duration(10, "seconds"))
+
+      val genStateRepo = inject[UserRecommendationGenerationStateRepo]
+      val re = db.readOnlyMaster { implicit session =>
+        genStateRepo.getByUserId(Id[User](42))
       }
-    }
-
-    "get adhoc recommendations" in {
-      withDb(modules: _*) { implicit injector =>
-        val repo = inject[UriRecommendationRepo]
-        db.readWrite { implicit s =>
-          val recs = setupRecs()
-          repo.save(recs(0))
-          repo.save(recs(1))
-          repo.save(recs(2))
-        }
-
-        val commander = inject[RecommendationGenerationCommander]
-        val result1 = commander.getAdHocRecommendations(Id[User](42), 2, UriRecommendationScores(socialScore = Some(0.5f)))
-        val recs1 = Await.result(result1, Duration(10, "seconds"))
-        recs1(0).userId === Some(Id[User](42))
-        recs1(0).score === 0.75f
-        recs1(1).score === 0.005f
-
-        val result2 = commander.getAdHocRecommendations(Id[User](42), 1, UriRecommendationScores())
-        val recs2 = Await.result(result2, Duration(10, "seconds"))
-        recs2(0).score === 48.0f
-
-      }
-    }
-
-    "pre-compute recommendations" in {
-      withDb(modules: _*) { implicit injector =>
-        val (user1, user2, shoebox) = setup()
-        val user1Keeps = makeKeeps(user1, 5, shoebox)
-        val user2Keeps = makeKeeps(user2, 5, shoebox)
-        val seedCommander = inject[SeedIngestionCommander]
-        val seedItemRepo = inject[RawSeedItemRepo]
-        val commander = inject[RecommendationGenerationCommander]
-        Await.result(seedCommander.ingestAllKeeps(), Duration(10, "seconds"))
-        db.readWrite { implicit session => seedItemRepo.assignSequenceNumbers(1000) }
-        commander.precomputeRecommendations()
-        val futUnit = commander.precomputeRecommendations()
-        Await.result(futUnit, Duration(10, "seconds"))
-        val result = commander.getTopRecommendations(Id[User](42), 1)
-        val recs = Await.result(result, Duration(10, "seconds"))
-        recs.size === 0
-      }
-    }
-
-    "pre-compute public feeds" in {
-      withDb(modules: _*) { implicit injector =>
-        val (user1, user2, shoebox) = setup()
-        val user1Keeps = makeKeeps(user1, 5, shoebox)
-        val user2Keeps = makeKeeps(user2, 5, shoebox)
-        val seedCommander = inject[SeedIngestionCommander]
-        val seedItemRepo = inject[RawSeedItemRepo]
-        val commander = inject[RecommendationGenerationCommander]
-        Await.result(seedCommander.ingestAllKeeps(), Duration(10, "seconds"))
-        db.readWrite { implicit session => seedItemRepo.assignSequenceNumbers(1000) }
-        val futUnit = commander.precomputePublicFeeds()
-        Await.result(futUnit, Duration(10, "seconds"))
-        val result = commander.getPublicFeeds(5)
-        val feeds = Await.result(result, Duration(10, "seconds"))
-        feeds.size === 0
-      }
-    }
-
-    "re-compute recos for user" in {
-      withDb(modules: _*) { implicit injector =>
-
-        val (user1, user2, shoebox) = setup()
-        val user1Keeps = makeKeeps(user1, 5, shoebox)
-        val user2Keeps = makeKeeps(user2, 5, shoebox)
-        val seedCommander = inject[SeedIngestionCommander]
-        val seedItemRepo = inject[RawSeedItemRepo]
-        val commander = inject[RecommendationGenerationCommander]
-        Await.result(seedCommander.ingestAllKeeps(), Duration(10, "seconds"))
-        db.readWrite { implicit session => seedItemRepo.assignSequenceNumbers(1000) }
-        commander.precomputeRecommendations()
-        val futUnit = commander.precomputeRecommendations()
-        Await.result(futUnit, Duration(10, "seconds"))
-        val result = commander.getTopRecommendations(Id[User](42), 1)
-        val recs = Await.result(result, Duration(10, "seconds"))
-        recs.size === 0
-
-        val resetFut = commander.resetUser(Id[User](42))
-        Await.result(resetFut, Duration(10, "seconds"))
-
-        val genStateRepo = inject[UserRecommendationGenerationStateRepo]
-        val re = db.readOnlyMaster { implicit session =>
-          genStateRepo.getByUserId(Id[User](42))
-        }
-        re.get.seq === SequenceNumber.ZERO
-        val newResult = commander.getTopRecommendations(Id[User](42), 1)
-        val newRecs = Await.result(newResult, Duration(10, "seconds"))
-        newRecs.size === 0
-      }
+      re.get.seq === SequenceNumber.ZERO
+      val newResult = commander.getTopRecommendations(Id[User](42), 1)
+      val newRecs = Await.result(newResult, Duration(10, "seconds"))
+      newRecs.size === 0
     }
   }
 }
